Fix: AlcoholListResponse 자료형 수정 (#14)

+++ PuppyMode/Responses/AlcoholListResponse.swift
@@ -23,7 +23,7 @@ struct AlcoholCategoryDetail: Codable {
 struct AlcoholListItem: Codable {
     let itemId: Int
     let itemName: String
-    let alcoholPercentage: Int
+    let alcoholPercentage: Double
     let volumeMl: Int?
     let imageUrl: String?
 }
